make cache reuse test more platform independent, minor exception handling improvements
